Backend/requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
pydantic==1.10.11
requests==2.31.0
beautifulsoup4==4.12.2
readability-lxml==0.8.1
sentence-transformers==2.2.2
faiss-cpu>=1.7.2
numpy==1.26.5
python-dotenv==1.2.1
openai==2.7.1
tqdm==4.66.1
pandas
pypdf2
langchain_community
langchain_openai
langchain
langchain.schema